+++ crypto/encryption/cs_paillier_test.go
@@ -15,7 +15,7 @@
  *
  */
 
-package test
+package encryption
 
 import (
 	"math/big"
@@ -25,23 +25,10 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/xlab-si/emmy/config"
 	"github.com/xlab-si/emmy/crypto/common"
-	"github.com/xlab-si/emmy/crypto/encryption"
 )
 
-func TestPaillier(t *testing.T) {
-	paillier := encryption.NewPaillier(1024)
-	pubKey := paillier.GetPubKey()
-
-	m := common.GetRandomInt(big.NewInt(123412341234123))
-	pubPaillier := encryption.NewPubPaillier(pubKey)
-	c, _ := pubPaillier.Encrypt(m)
-	p, _ := paillier.Decrypt(c)
-
-	assert.Equal(t, m, p, "Paillier encryption/decryption does not work correctly")
-}
-
 func TestCSPaillier(t *testing.T) {
-	secParams := encryption.CSPaillierSecParams{
+	secParams := CSPaillierSecParams{
 		L:        512,
 		RoLength: 160,
 		K:        158,
@@ -52,17 +39,17 @@ func TestCSPaillier(t *testing.T) {
 	secKeyPath := filepath.Join(dir, "cspaillierseckey.txt")
 	pubKeyPath := filepath.Join(dir, "cspaillierpubkey.txt")
 
-	cspaillier := encryption.NewCSPaillier(&secParams)
+	cspaillier := NewCSPaillier(&secParams)
 	cspaillier.StoreSecKey(secKeyPath)
 	cspaillier.StorePubKey(pubKeyPath)
 
-	cspaillierPub, _ := encryption.NewCSPaillierFromPubKeyFile(pubKeyPath)
+	cspaillierPub, _ := NewCSPaillierFromPubKeyFile(pubKeyPath)
 
 	m := common.GetRandomInt(big.NewInt(8685849))
 	label := common.GetRandomInt(big.NewInt(340002223232))
 	u, e, v, _ := cspaillierPub.Encrypt(m, label)
 
-	cspaillierSec, _ := encryption.NewCSPaillierFromSecKey(secKeyPath)
+	cspaillierSec, _ := NewCSPaillierFromSecKey(secKeyPath)
 	p, _ := cspaillierSec.Decrypt(u, e, v, label)
 
 	assert.Equal(t, m, p, "Camenisch-Shoup modified Paillier encryption/decryption does not work correctly")
